Compute sin(acos(x)) as sqrt(1-dot*dot), which is faster.
Notice that despite the subtraction "1-dot*dot" there is no significant
loss of accuracy here, even for small angles where "dot" is near 1. This is
because our knowledge of the angle from "dot (i.e. cos(theta)) is already
limited by the floating representation error near 1, which is 2^-23 (1.1e-7)
for single and 2^-52 (2.2e-16) for double precision.

The following program tabulates the relative error of the two formulas,
evaluated in single precision, compared to evaluating the original formula
in double precsion:

https://godbolt.org/z/qrzWcd43c